requirements.txt
Flask==3.0.3
Werkzeug>=3.0.0
openai
python-dotenv
openpyxl
gunicorn
psycopg2-binary
